Move all incremental Java performance tests to one class
